refactor: rename projects to Projects and clean up css
