+++ client/src/Projects/Detail.tsx
